+++ htdocs/components/ActivityComponent/subComponents/ActivityEditComponent.js
@@ -132,6 +132,7 @@ class ActivityEditComponent extends MainComponent {
         >
         </CallbackModal>
     }
+
     componentDidMount() {
         this.loadCallActivity(params.get("callActivityID"));
 
@@ -1913,7 +1914,7 @@ class ActivityEditComponent extends MainComponent {
             return '';
         }
         return <AssetListSelectorComponent
-            noAssetReason={data.emptyAssetReason}
+            emptyAssetReason={data.emptyAssetReason}
             assetName={data.assetName}
             assetTitle={data.assetTitle}
             customerId={data.customerId}