+++ htdocs/components/AssetPickerComponent/AssetPickerComponent.js
@@ -2,11 +2,11 @@ import ReactDOM from "react-dom";
 import React from "react";
 import AssetListSelectorComponent from "../shared/AssetListSelectorComponent/AssetListSelectorComponent";
 
-document.AssetPickerComponentRender = (elementId, customerId, noAssetReason, assetName,assetTitle, onChange) => {
+document.AssetPickerComponentRender = (elementId, customerId, emptyAssetReason, assetName, assetTitle, onChange) => {
     const domContainer = document.querySelector(elementId);
     ReactDOM.render(React.createElement(AssetListSelectorComponent, {
         customerId,
-        noAssetReason,
+        emptyAssetReason,
         assetName,
         assetTitle,
         onChange